refactor: name the JSON-repr model base and lift the benchmark kernels out
MyBaseModel said nothing about what it provides -- indented-JSON rendering for repr, str and show(), and nothing else -- so it becomes JsonReprModel. It is private to the core and exported nowhere, so no public name changes. The 44 flops benchmark kernels move from inside the suite builder to their own module: each closes over nothing and takes every input as an argument, so where they are defined does not change the machine code numba emits (confirmed by an interleaved before/after measurement), and lifting them takes the builder from 629 lines under a C901 exemption to a flat 244-line registry with the gate back on.
