+++ src/counted_float/_core/benchmarking/counted_float/_counted_float_benchmark.py
@@ -1,13 +1,13 @@
 from counted_float._core.benchmarking.micro import MicroBenchmark
 from counted_float._core.counting import CountedFloat
-from counted_float._core.models import MyBaseModel
+from counted_float._core.models import JsonReprModel
 from counted_float._core.utils import format_time_duration
 
 
 # =================================================================================================
 #  Result class
 # =================================================================================================
-class CountedFloatBenchmarkResults(MyBaseModel):
+class CountedFloatBenchmarkResults(JsonReprModel):
     float_time_nsec: float
     counted_float_time_nsec: float
