+++ tests/models/test_flops_benchmark_results.py
@@ -7,7 +7,7 @@ if TYPE_CHECKING:
 
 
 def test_flops_benchmark_results_show():
-    """Minimal test to check if MyBaseModel.show() at least does not raise exceptions."""
+    """Minimal test to check if JsonReprModel.show() at least does not raise exceptions."""
 
     # --- arrange -----------------------------------------
     flops_benchmark_results: FlopsBenchmarkResults = list(BuiltInData.benchmarks().values()).pop()
